+++ src/main/java/subway/ui/controller/station/DeleteStationController.java
@@ -1,4 +1,4 @@
-package subway.ui.station;
+package subway.ui.controller.station;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
